hiding websocket initalization behind authentication
